UwezoFX - 4/requirements.txt
MetaTrader5
pandas
numpy
scipy
scikit-learn
Pillow
supabase
python-dotenv
sib-api-v3-sdk
africastalking
cryptography
requests
myfxbook